31. merge request şerefine

See merge request sulinos/devel/inary!33

+++ Makefile
@@ -1,5 +1,5 @@
 DESTDIR=/
-all: build install
+all: build
 
 clean:
 	`find | grep pycache | sed 's/^/rm -rf /g'`
@@ -9,4 +9,4 @@ build:
 	python3 setup.py build
 install:
 	python3 setup.py install --install-lib=${DESTDIR}/usr/lib/sulin --root=${DESTDIR}
-	ln -s inary-cli ${DESTDIR}/usr/bin/inary || true
+	[ -f ${DESTDIR}/usr/bin/inary ] || ln -s inary-cli ${DESTDIR}/usr/bin/inary || true
